requirements.txt
Django==2.1
django-bootstrap4==0.0.7
django-filter==2.0.0
django-select2==6.3.1
Jinja2==2.10
psycopg2-binary==2.7.7
